Improve Sampler module
